util: replace float with double in Vector
- Justin C

+++ src/main/include/util/vector.h
@@ -2,18 +2,18 @@
 
 class Vector {
 public:
-    Vector(float x, float y, float z) : x(x), y(y), z(z) {}
+    Vector(double x, double y, double z) : x(x), y(y), z(z) {}
     Vector() { // Heehoo Default Constructor
         x = 0;
         y = 0;
         z = 0;
     }
 
-    float dot(Vector & rhs) {
+    double dot(Vector & rhs) {
         return (x * rhs.x) + (y * rhs.y) + (z * rhs.z);
     }
 
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 };
